refactor: extract fetch tokens logic to custom hook
