+++ hooks/index.ts
@@ -1,3 +1,4 @@
 export {default as useHandleDestroyAnimated} from './useHandleDestroyAnimated';
 export {default as useMintAndTransfer} from './useMintAndTransfer';
 export {default as useCreateToken} from './useCreateToken';
+export {default as useFetchTokens} from './useFetchTokens';
